[#171405467] Extract getApiClient() method into utils/apim module
