+++ utils/apim.ts
@@ -1,0 +1,31 @@
+import { ApiManagementClient } from "@azure/arm-apimanagement";
+import * as msRestNodeAuth from "@azure/ms-rest-nodeauth";
+import { toError } from "fp-ts/lib/Either";
+import { TaskEither, tryCatch } from "fp-ts/lib/TaskEither";
+
+export interface IServicePrincipalCreds {
+  readonly clientId: string;
+  readonly secret: string;
+  readonly tenantId: string;
+}
+
+export interface IAzureApimConfig {
+  readonly subscriptionId: string;
+  readonly apimResourceGroup: string;
+  readonly apim: string;
+}
+
+export function getApiClient(
+  servicePrincipalCreds: IServicePrincipalCreds,
+  subscriptionId: string
+): TaskEither<Error, ApiManagementClient> {
+  return tryCatch(
+    () =>
+      msRestNodeAuth.loginWithServicePrincipalSecret(
+        servicePrincipalCreds.clientId,
+        servicePrincipalCreds.secret,
+        servicePrincipalCreds.tenantId
+      ),
+    toError
+  ).map(credentials => new ApiManagementClient(credentials, subscriptionId));
+}
